refactor: history storage handled by databaseIO
