+++ app/src/androidTest/java/ch/epfl/sdp/storage/PositionHistoryManagerTest.java
@@ -13,9 +13,12 @@ import java.util.SortedMap;
 
 import ch.epfl.sdp.contamination.Carrier;
 import ch.epfl.sdp.contamination.Layman;
+import ch.epfl.sdp.contamination.databaseIO.ConcreteDataReceiver;
 import ch.epfl.sdp.contamination.databaseIO.ConcreteDataSender;
+import ch.epfl.sdp.contamination.databaseIO.DataReceiver;
 import ch.epfl.sdp.contamination.databaseIO.DataSender;
 import ch.epfl.sdp.contamination.databaseIO.GridFirestoreInteractor;
+import ch.epfl.sdp.contamination.databaseIO.PositionHistoryManager;
 import ch.epfl.sdp.identity.fragment.AccountFragment;
 
 import static ch.epfl.sdp.TestTools.newLoc;
@@ -24,15 +27,13 @@ import static org.junit.Assert.assertTrue;
 
 public class PositionHistoryManagerTest {
     private DataSender sender;
+    private DataReceiver receiver;
     @Before
     public void miniInit(){
-        sender = new ConcreteDataSender(new GridFirestoreInteractor());
+        GridFirestoreInteractor grid = new GridFirestoreInteractor();
+        sender = new ConcreteDataSender(grid);
+        receiver = new ConcreteDataReceiver(grid);
         AccountFragment.IN_TEST = true;
-        PositionHistoryManager.delete();
-    }
-    @After
-    public void release(){
-        PositionHistoryManager.delete();
     }
 
     @Test
@@ -41,7 +42,7 @@ public class PositionHistoryManagerTest {
         Layman me = new Layman(Carrier.InfectionStatus.HEALTHY);
         sender.registerLocation(me, newLoc(2, 2), new Date(System.currentTimeMillis() - 1 - PRESYMPTOMATIC_CONTAGION_TIME));
         sender.registerLocation(me, newLoc(1, 1), new Date(System.currentTimeMillis()));
-        SortedMap<Date, Location> res = PositionHistoryManager.getLastPositions();
+        SortedMap<Date, Location> res = receiver.getLastPositions();
         Collection<Location> val = res.values();
         Iterator<Location> it = val.iterator();
         assertTrue(val.size() == 1);
